requirements.txt
python-dotenv
openai
pdfminer.six
python-docx
flask
flask_cors
PyYAML
cloudinary==1.36.0